requirements.txt
# Core
anthropic>=0.18.0
openai>=1.12.0
httpx>=0.27.0
python-dotenv>=1.0.0

# API (for n8n / HTTP trigger)
fastapi>=0.115.0
uvicorn[standard]>=0.32.0

# Database
psycopg2-binary>=2.9.9
supabase>=2.3.0

# YouTube & media
google-api-python-client>=2.111.0
yt-dlp>=2024.1.0

# Transcription
deepgram-sdk>=3.0.0

# Search
meilisearch>=0.31.0
